feat: 편집 로그 상세에서 모바일에서 버튼이 보이게 설정
