+++ client/src/components/DocumentPage.tsx
@@ -1,12 +1,10 @@
 import React, { useEffect } from 'react';
 import { useNavigate, useParams } from 'react-router-dom';
 import useGetDocumentByTitle from '@api/get/useGetDocumentByTitle';
-import { useQueryClient } from '@tanstack/react-query';
-import URLS from '@constants/urls';
 import DocumentHeader from './DocumentHeader';
 import DocumentContents from './DocumentContents';
 import DocumentFooter from './DocumentFooter';
-import Button from './Button';
+import MobileDocumentHeader from './MobileDocumentHeader';
 
 interface DocumentPageProps {
   daemoon?: string;
@@ -24,35 +22,9 @@ const DocumentPage = ({ daemoon }: DocumentPageProps) => {
     }
   }, [daemoon]);
 
-  const queryClient = useQueryClient();
-
-  const refreshData = () => {
-    queryClient.removeQueries();
-    navigate('');
-  };
-
-  const goPostPage = () => {
-    navigate(URLS.POST);
-  };
-
-  const goEditPage = () => {
-    navigate(URLS.EDIT, { state: docs });
-  };
-
-  const goLogsPage = () => {
-    navigate(URLS.LOGS, { state: docs });
-  };
-
   return (
     <>
-      <div className="md:hidden">
-        <fieldset className="flex gap-2 max-md:w-full max-md:justify-center">
-          <Button style="tertiary" size="xs" text="새로고침" onClick={refreshData} />
-          <Button style="tertiary" size="xs" text="편집하기" onClick={goEditPage} />
-          <Button style="tertiary" size="xs" text="편집로그" onClick={goLogsPage} />
-          <Button style="primary" size="xs" text="작성하기" onClick={goPostPage} />
-        </fieldset>
-      </div>
+      <MobileDocumentHeader docs={docs} />
       <div className="flex flex-col gap-6 w-full h-fit min-h-[864px] bg-white border-primary-100 border-solid border rounded-xl p-8 max-md:p-4 max-md:gap-2">
         <DocumentHeader wiki={docs} />
         <DocumentContents contents={docs.contents} />
