+++ client/src/components/LogContent.tsx
@@ -4,6 +4,7 @@ import { useParams } from 'react-router-dom';
 import DocumentHeader from './DocumentHeader';
 import DocumentContents from './DocumentContents';
 import DocumentFooter from './DocumentFooter';
+import MobileDocumentHeader from './MobileDocumentHeader';
 
 const LogContent = () => {
   const { logId } = useParams();
@@ -11,6 +12,7 @@ const LogContent = () => {
 
   return (
     <>
+      <MobileDocumentHeader docs={document} />
       <div className="flex flex-col gap-6 w-full h-fit bg-white border-primary-100 border-solid border rounded-xl p-8 max-[768px]:p-4 max-[768px]:gap-2 ">
         <DocumentHeader wiki={{ documentId: document.logId, ...document }} />
         <DocumentContents contents={document.contents} />
